requirements.txt
pynput>=1.7.0
PyYAML>=6.0